--- a/prometheus/value.go
+++ b/prometheus/value.go
@@ -46,6 +46,9 @@ type value struct {
 	// operations.  http://golang.org/pkg/sync/atomic/#pkg-note-BUG
 	valBits uint64
 
+	// valInt is used to store values that are exact integers
+	valInt int64
+
 	selfCollector
 
 	desc       *Desc
@@ -82,15 +85,25 @@ func (v *value) SetToCurrentTime() {
 	v.Set(float64(time.Now().UnixNano()) / 1e9)
 }
 
+// add adjusts the underlying int64
+func (v *value) add(delta int64) {
+	atomic.AddInt64(&v.valInt, delta)
+}
+
 func (v *value) Inc() {
-	v.Add(1)
+	v.add(1)
 }
 
 func (v *value) Dec() {
-	v.Add(-1)
+	v.add(-1)
 }
 
 func (v *value) Add(val float64) {
+	if math.Trunc(val) == val {
+		atomic.AddInt64(&v.valInt, int64(val))
+		return
+	}
+
 	for {
 		oldBits := atomic.LoadUint64(&v.valBits)
 		newBits := math.Float64bits(math.Float64frombits(oldBits) + val)
@@ -105,7 +118,10 @@ func (v *value) Sub(val float64) {
 }
 
 func (v *value) Write(out *dto.Metric) error {
-	val := math.Float64frombits(atomic.LoadUint64(&v.valBits))
+	fval := math.Float64frombits(atomic.LoadUint64(&v.valBits))
+	ival := atomic.LoadInt64(&v.valInt)
+	val := fval + float64(ival)
+
 	return populateMetric(v.valType, val, v.labelPairs, out)
 }