We would like to use something in the prometheus ecosystem with comparable performance to `github.com/rcrowley/go-metrics`. It seems as though the use of [`atomic.CompareAndSwap`](https://github.com/prometheus/client_golang/blob/master/prometheus/value.go#L97) slows things down a bit, so we've made a counter that stores and uses an `int64` directly and leans on `atomic.*Int64` for locking.

This branch introduces the `FixedPrecisionCounter` type and I'd like some feedback on the approach generally (and name) before I plumb through the vector version of this counter.

As to how I benchmarked performance I used the following test file:

```go
package fpc

import (
        "testing"

        "github.com/prometheus/client_golang/prometheus"
        metrics "github.com/rcrowley/go-metrics"
)

var fpc prometheus.Counter
var p prometheus.Counter

func init() {
        fpc = prometheus.NewFixedPrecisionCounter(prometheus.Opts{
                Name: "foo",
                Help: "helpful message about foo",
        }, 0)
        p = prometheus.NewCounter(prometheus.CounterOpts{Name: "pfoo", Help: "pfoo help"})
        prometheus.MustRegister(fpc)
        prometheus.MustRegister(p)
}

func BenchmarkGoMetrics(b *testing.B) {
        c := metrics.NewCounter()
        b.RunParallel(func(pb *testing.PB) {
                for pb.Next() {
                        c.Inc(1)
                }
        })
}

func BenchmarkPromCounter(b *testing.B) {
        b.RunParallel(func(pb *testing.PB) {
                for pb.Next() {
                        p.Inc()
                }
        })
}

func BenchmarkFixedPrecisionCounter(b *testing.B) {
        b.RunParallel(func(pb *testing.PB) {
                for pb.Next() {
                        fpc.Inc()
                }
        })
}
```

We get the following performance numbers:

```
go test -cpu=64 -benchtime=5s -v -bench=. github.com/smcquay/fpc
goos: darwin
goarch: amd64
pkg: github.com/smcquay/fpc
BenchmarkGoMetrics-64                   500000000               13.2 ns/op
BenchmarkPromCounter-64                 100000000               62.1 ns/op
BenchmarkFixedPrecisionCounter-64       300000000               22.5 ns/op
PASS
ok      github.com/smcquay/fpc  23.261s
```