@brian-brazil 2017-12-06
Prometheus only supports float64 internally, so this won't work.