@beorn7 2017-12-06
@smcquay thanks for your contribution. I was tempted to add something like that long ago but then decided against it because of the additional cognitive load for the user for a performance improvement that is only relevant in very specific scenarios. Note that this had to be implemented for both counters and gauges (you incidentally implemented a fixed-precision gauge, not a counter), which adds a lot more top-level types to the already quite confusing package. And then users can easily get things wrong in subtle ways by picking the fixed-precision version when they should not have.

I would much prefer to first try our best to optimize within the bounds of the existing interfaces. For example, you could just count `Inc` calls in a separate integer and only add them to the final value in the `Write` call. `Add` calls with an effective integer (e.g. `Add(42.0)`) could use the same intermediate integer.