@smcquay 2017-12-14
@beorn7 Thanks for the feedback. I am going to work a branch that adds the optimization you've suggested, but I'm happy moving this implementation into a separate package where people can use it if they want this flavor of optimization. However I'd like to use this new implementation with the prometheus client vector implementations, however due to unexported types I cannot conveniently use this external type with `metricVec`.

What are your thoughts about exporting this type? I understand the desire to keep the documentation as narrow as possible for users, and perhaps this is why `metricVec` is unexported, but would you think it reasonable to expose this type so that I can make my own `*Vec` types?

This approach would allow users to create their own completely decoupled `prometheus.Metric` implementations and still gain the benefits of using those within this client ecosystem.

Thoughts?