@beorn7 2017-12-14
> What are your thoughts about exporting this type? I understand the desire to keep the documentation as narrow as possible for users, and perhaps this is why metricVec is unexported, but would you think it reasonable to expose this type so that I can make my own *Vec types?
>
> This approach would allow users to create their own completely decoupled prometheus.Metric implementations and still gain the benefits of using those within this client ecosystem.

I understand your need. And it is exactly the reason why `metricVec` used to be an exported type for quite some time. However, nobody ever used it, but I got feedback about how unwieldy the package is with the many exported type visible to the naive user (that only wants to do a few simple direct instrumentation and doesn't need most of them). So I started a multi-step effort to reduce the number of exported types. Which made `metricVec` the way it is.

Perhaps there is a way to create a separate package that would be some kind of implementation helper for developers like you. But it's not entirely trivial without creating circular package dependencies… I'd be delighted if you have an idea along these lines.