@smcquay 2017-12-14
@beorn7 --

> Perhaps there is a way to create a separate package that would be some kind of implementation helper for developers like you. But it's not entirely trivial without creating circular package dependencies… I'd be delighted if you have an idea along these lines.

I wouldn't mind taking a crack at it, but have concerns about doing the work to have that PR declined. My design goals for this lift would be to keep [top-level docs](https://godoc.org/github.com/prometheus/client_golang/prometheus) same size (or smaller) and expose the functionality I need immediately (probably just `MetricVec`) to demonstrate feasibility of the effort. 

Can you please validate that the approach works for you? Anything else you'd like me to keep an eye on as I do this refactor?