@smcquay 2017-12-14
I'm starting to gain an appreciation for the difficulty of the cyclic package dependency problem :)

A possible approach: how about we have all of the useful functionality exposed in a sub package, perhaps as it was earlier. Then we could expose a filtered view on that large package by having the `github.com/prometheus/client_golang/prometheus` package simply reference (and document) the types within the package that has the superset of exported objects? Something of the form:

```go
package prometheus

// ...

// A Metric models a single sample value with its meta data being exported to
// Prometheus. Implementations of Metric in this package are Gauge, Counter,
// Histogram, Summary, and Untyped.
type Metric prometheus.Metric

// ...

```

I'm guessing you've considered this approach already and we'd be trading the ability to use types externally (as I'd like to do with `MetricVec`) for additional work of duplicating documentation.