@beorn7 2017-12-14
We might go down that way one day. However, that would result in a major refactoring, which might be more appropriate once the changes planned for v0.10 are in, which are by itself quite invasive.

My experience so far is that there are a _lot_ of sometimes contradicting ideas and requirements users of this package have. It is hard to optimize for satisfying as many of them as possible. I initially valued the option to implement own `Metric` and `MetricVec` types pretty highly but then it was never needed by anybody (until now). So I changed directions and weighed other properties stronger.

Given the huge backlog of more concrete demand for other features, I'm hesitant to commit to a larger restructuring to satisfy a relatively rare requirement at this point. It even looks that most of the gain for your use case can be had by implementing the transparent optimizations I have contemplated above.

Perhaps it's really the easiest if you keep your specialized counters in a friendly fork of this repo for now and we take a raincheck a bit further down the path.

In any case, I'd be happy to review and accept any under-the-hood optimizations that are transparent for the user (see above for ideas).