@smcquay 2017-12-14
@beorn7 I agree with your assessment. Thanks for helping me reason through this. I'm taking a whack at an under-the-hood optimization.