@smcquay 2017-12-14
Alright had a whack at it and it seems to be within striking distance:

```
 ⌘ go test -cpu=64 -benchtime=5s -v -bench=. github.com/smcquay/counters/bench
goos: darwin
goarch: amd64
pkg: github.com/smcquay/counters/bench
BenchmarkGoMetrics-64                   500000000               15.5 ns/op
BenchmarkPromCounter-64                 500000000               18.6 ns/op
BenchmarkFixedPrecisionCounter-64       300000000               24.5 ns/op
PASS
ok      github.com/smcquay/counters/bench       30.160s
```