@smcquay 2017-12-15
Implemented your review feedback and am considering how to attack the concurrent `Set`/`Write` problem without introducing locks. The naive approach that comes to mind is to use a `RWMutex` in both functions, but that seems heavy-handed. Please do let me know if you come up with a better idea.